+++ .changeset/spicy-grapes-notice.md
@@ -1,0 +1,6 @@
+---
+'@project44-manifest/react': minor
+'@project44-manifest/react-dropdown': minor
+---
+
+Adding dropdown package
