+++ packages/react/components/dropdown/README.md
@@ -1,0 +1,4 @@
+# @project44-manifest/react-dropdown
+
+This package is part of [Manifest Design System](https://github.com/project44/manifest). Please see
+the repo for more details.
