+++ packages/react/components/dropdown/moon.yml
@@ -1,0 +1,31 @@
+type: 'library'
+language: 'typescript'
+
+workspace:
+  inheritedTasks:
+    rename:
+      buildPackage: 'build'
+
+tasks:
+  build:
+    outputs:
+      - 'esm'
+      - 'lib'
+
+dependsOn:
+  - id: 'button'
+    scope: 'development'
+  - id: 'icons'
+    scope: 'development'
+  - id: 'overlay'
+    scope: 'production'
+  - id: 'popover'
+    scope: 'production'
+  - id: 'styles'
+    scope: 'production'
+  - id: 'types'
+    scope: 'development'
+  - id: 'typography'
+    scope: 'production'
+  - id: 'utils'
+    scope: 'production'
