+++ packages/react/components/dropdown/package.json
@@ -1,0 +1,73 @@
+{
+  "name": "@project44-manifest/react-dropdown",
+  "version": "0.0.0",
+  "description": "Manifest Design System react dropdown component",
+  "license": "MIT",
+  "author": "project44",
+  "keywords": [
+    "manifest",
+    "design",
+    "system",
+    "react",
+    "components"
+  ],
+  "sideEffects": false,
+  "main": "./lib/index.js",
+  "module": "./esm/index.js",
+  "types": "./dts/index.d.ts",
+  "files": [
+    "dts/**/*.d.ts",
+    "esm/**/*.{js,map}",
+    "lib/**/*.{js,map}",
+    "src/**/*.{ts,tsx,json}"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:project-44/manifest.git",
+    "directory": "packages/dropdown"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "peerDependencies": {
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
+  },
+  "dependencies": {
+    "@project44-manifest/react-overlay": "^0.0.0",
+    "@project44-manifest/react-popover": "^0.1.2",
+    "@project44-manifest/react-styles": "^1.0.1",
+    "@project44-manifest/react-typography": "^1.0.1",
+    "@project44-manifest/react-utils": "^0.2.3",
+    "@react-aria/focus": "^3.10.1",
+    "@react-aria/interactions": "^3.13.1",
+    "@react-aria/menu": "^3.7.1",
+    "@react-aria/separator": "^3.2.6",
+    "@react-stately/collections": "^3.5.1",
+    "@react-stately/menu": "^3.4.4",
+    "@react-stately/tree": "^3.4.1"
+  },
+  "devDependencies": {
+    "@project44-manifest/react-button": "^1.0.2",
+    "@project44-manifest/react-icons": "^0.1.1",
+    "@project44-manifest/react-types": "^0.2.1",
+    "@react-types/shared": "^3.16.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
+  },
+  "packemon": {
+    "platform": "browser"
+  },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "types": "./dts/index.d.ts",
+      "browser": {
+        "module": "./esm/index.js",
+        "import": "./esm/index.js",
+        "default": "./lib/index.js"
+      },
+      "default": "./lib/index.js"
+    }
+  }
+}
