+++ packages/react/components/dropdown/src/Dropdown.context.ts
@@ -1,0 +1,12 @@
+import { createContext } from '@project44-manifest/react-utils';
+
+export interface DropdownContext {
+  closeOnSelect?: boolean;
+  menuProps: React.HTMLAttributes<HTMLUListElement>;
+  menuRef: React.RefObject<HTMLUListElement>;
+  onClose?: () => void;
+}
+
+export const [DropdownProvider, useDropdownContext] = createContext<DropdownContext>({
+  name: 'DropdownContext',
+});
