+++ packages/react/components/dropdown/src/Dropdown.styles.ts
@@ -1,0 +1,5 @@
+import { styled } from '@project44-manifest/react-styles';
+
+export const StyledDropdown = styled('div', {
+  boxSizing: 'border-box',
+});
