+++ packages/react/components/dropdown/src/DropdownItem/index.ts
@@ -1,8 +1,9 @@
 import { Item } from '@react-stately/collections';
-import type { DropdownItemProps } from './DropdownItem';
+import type { DropdownItemProps } from './DropdownItem.types';
 
 export const DropdownItem = Item as (
   props: Omit<DropdownItemProps, 'item' | 'onAction' | 'state'>,
 ) => JSX.Element;
 
-export { type DropdownItemProps, DropdownItem as DROPDOWN_ITEM } from './DropdownItem';
+export { DropdownItem as DROPDOWN_ITEM } from './DropdownItem';
+export type { DropdownItemProps };
