+++ packages/react/components/dropdown/src/DropdownMenu/DropdownMenu.types.ts
@@ -1,0 +1,11 @@
+import type { AriaMenuProps } from '@react-types/menu';
+import type { CSS } from '@project44-manifest/react-styles';
+
+export type DropdownMenuElement = 'ul';
+
+export interface DropdownMenuProps extends AriaMenuProps<object> {
+  /**
+   * Theme aware style object
+   */
+  css?: CSS;
+}
