+++ packages/react/components/dropdown/src/DropdownSection/index.ts
@@ -1,0 +1,7 @@
+import { Section } from '@react-stately/collections';
+import type { DropdownSectionProps } from './DropdownSection.types';
+
+export const DropdownSection = Section as (props: DropdownSectionProps) => JSX.Element;
+
+export { DropdownSection as DROPDOWN_SECTION } from './DropdownSection';
+export type { DropdownSectionProps };
