+++ packages/react/components/dropdown/src/index.ts
@@ -1,0 +1,5 @@
+export * from './Dropdown';
+export * from './Dropdown.types';
+export * from './DropdownItem';
+export * from './DropdownMenu';
+export * from './DropdownSection';
