--- packages/react/core/src/components/Dropdown/Dropdown.context.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-
-export interface DropdownContext {
-  closeOnSelect?: boolean;
-  menuProps: React.HTMLAttributes<HTMLUListElement>;
-  menuRef: React.RefObject<HTMLUListElement>;
-  onClose?: () => void;
-}
-
-export const DropdownContext = React.createContext<DropdownContext | null>(null);
-
-export const useDropdownContext = () => React.useContext(DropdownContext);
