--- packages/react/core/src/components/Dropdown/index.ts
@@ -1,2 +1,0 @@
-export * from './Dropdown';
-export * from './Dropdown.context';
