+++ packages/react/core/src/index.ts
@@ -18,14 +18,6 @@ export type { DatePickerProps } from './components/DatePicker';
 export { DatePicker } from './components/DatePicker';
 export type { DateRangePickerProps } from './components/DateRangePicker';
 export { DateRangePicker } from './components/DateRangePicker';
-export type { DropdownProps } from './components/Dropdown';
-export { Dropdown } from './components/Dropdown';
-export type { DropdownItemProps } from './components/DropdownItem';
-export { DropdownItem } from './components/DropdownItem';
-export type { DropdownMenuProps } from './components/DropdownMenu';
-export { DropdownMenu } from './components/DropdownMenu';
-export type { DropdownSectionProps } from './components/DropdownSection';
-export { DropdownSection } from './components/DropdownSection';
 export type { FlexProps } from './components/Flex';
 export { Flex } from './components/Flex';
 export type { FormControlProps } from './components/FormControl';
